refactor : 더미데이터 폴더명 수정
